requirements.txt
asgiref==3.3.1
backcall==0.2.0
decorator==4.4.2
Django==3.1.5
djangorestframework==3.12.2
ipython==7.19.0
ipython-genutils==0.2.0
jedi==0.18.0
parso==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.10
ptyprocess==0.7.0
Pygments==2.7.4
pytz==2020.5
sqlparse==0.4.1
traitlets==5.0.5
wcwidth==0.2.5
